backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.20
pdfplumber==0.11.4
pdf2image==1.17.0
Pillow==11.1.0
anthropic==0.43.0
python-dotenv==1.0.1
aiofiles==24.1.0
alembic==1.14.0
